# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: MDM API
  version: 0.3.0
  summary: セルフホストの動画管理システム API
  description: |
    取り込み（スキャン）・一覧・再生・題名検索までを公開する。
    認証は掛けていない（Phase 3 の範囲）。信頼できるネットワークの内側で使う前提。
servers:
  - url: /
    description: アプリケーション自身が配信する
paths:
...
  /api/videos:
    get:
      operationId: listVideos
      summary: 動画の一覧を返す
      description: |
        カーソル方式でページングする。`query` を与えると題名の部分一致で絞り込む。
        `total` は絞り込み後の総件数で、ページングとは独立に返る。
      parameters:
        - name: query
          in: query
          description: 題名の部分一致。1文字から指定できる
          schema:
            type: string
            maxLength: 100
        - name: sort
          in: query
          description: 並び順
          schema:
            $ref: "#/components/schemas/VideoSort"
        - name: cursor
          in: query
          description: 前回の応答が返した `nextCursor`。中身は不透明で、解釈しない
          schema:
            type: string
        - name: limit
          in: query
          description: 1ページの件数
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 60
      responses:
        "200":
          description: 一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoPage"
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /api/videos/{id}:
    get:
      operationId: getVideo
      summary: 動画1件の詳細を返す
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "200":
          description: 詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/scans:
    post:
      operationId: startScan
      summary: 取り込みを開始する
      description: |
        すでに実行中の場合は、新しく始めずに実行中のスキャンを返す。
        応答は即座に返り、取り込みは背後で進む。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              maxProperties: 0
      responses:
        "202":
          description: 開始した、または実行中のスキャン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "409":
          description: メディアフォルダが未設定
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/media-folders:
    get:
      operationId: listMediaFolders
      summary: 登録済みメディアフォルダを返す
      responses:
        "200":
          description: id昇順のメディアフォルダ一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaFolder"
    post:
      operationId: createMediaFolder
      summary: メディアフォルダを1件追加する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaFolderRequest"
      responses:
        "201":
          description: 作成したメディアフォルダ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFolder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/media-folders/{id}:
    put:
      operationId: updateMediaFolder
      summary: メディアフォルダ1件のpathを変更する
      parameters:
        - $ref: "#/components/parameters/MediaFolderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMediaFolderRequest"
      responses:
        "200":
          description: 更新後のメディアフォルダ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFolder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteMediaFolder
      summary: メディアフォルダを1件削除する
      parameters:
        - $ref: "#/components/parameters/MediaFolderId"
        - name: version
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: 削除した
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
    VideoId:
      name: id
      in: path
      required: true
      description: 動画の識別子
      schema:
        type: integer
        format: int64
        minimum: 1
    MediaFolderId:
      name: id
      in: path
      required: true
      description: メディアフォルダの識別子
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    NotFound:
      description: 対象が存在しない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidRequest:
      description: 要求の形式が不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: same-originでない操作
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 現在の状態と競合する操作
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    MediaFolder:
      type: object
      required: [id, path, version, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        path:
          type: string
        version:
          type: integer
          format: int64
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateMediaFolderRequest:
      type: object
      required: [path]
      additionalProperties: false
      properties:
        path:
          type: string
          minLength: 1
    UpdateMediaFolderRequest:
      type: object
      required: [path, version]
      additionalProperties: false
      properties:
        path:
          type: string
          minLength: 1
        version:
          type: integer
          format: int64
          minimum: 1
...
    VideoSort:
      type: string
      enum: [addedDesc, titleAsc]
      default: addedDesc
      description: addedDesc = 追加が新しい順、titleAsc = 題名順
    VideoPage:
      type: object
      required: [items, total]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        total:
          type: integer
          description: 絞り込み後の総件数
        nextCursor:
          type: string
          description: 次のページの取得に渡す。これ以上無い場合は省略される
...
    Video:
      type: object
      required: [id, title, sizeBytes, addedAt, playable, probeState, thumbnailState, previewState]
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          description: 拡張子を除いたファイル名
        sizeBytes:
          type: integer
          format: int64
        addedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
          format: int64
          description: 尺。解析前・取得不能の場合は省略される
        width:
          type: integer
        height:
          type: integer
        container:
          type: string
          examples: [mp4, webm, mkv]
        videoCodec:
          type: string
          examples: [h264, vp9]
        audioCodec:
          type: string
          examples: [aac, opus]
        playable:
          type: boolean
          description: ブラウザでそのまま再生できると判定されたか
        unplayableReason:
          type: string
          enum: [container, video_codec, audio_codec]
          description: playable = false の理由。判定前は省略される
        probeState:
          type: string
          enum: [pending, done, failed]
        probeError:
          type: string
          description: probeState = failed のときの理由
        thumbnailState:
          type: string
          enum: [pending, done, failed]
        previewState:
          type: string
          enum: [pending, done, failed]
        thumbnailUrl:
          type: string
          description: thumbnailState = done のときだけ入る
        previewUrl:
          type: string
          description: previewState = done かつ保存済み asset が配信可能なときだけ入る版付き URL
        seekThumbnailUrl:
          type: string
          description: probeState = done かつ正のdurationMsを持つときだけ入る版付き基底URL
        progress:
          $ref: "#/components/schemas/Progress"
        location:
          $ref: "#/components/schemas/VideoLocation"
        seekThumbnailState:
          type: string
          enum: [pending, done, failed]
          description: |
            シーク用プレビューの状態。GET /api/videos/{id} の応答にだけ入り、
            seekThumbnailUrl と同じ条件のときだけ入る。done = 置き場がある、
            pending = 置き場が無く thumbnailState = pending かサムネイルのジョブが
            queued・running、failed = それ以外
    VideoLocation:
      type: object
      required: [path, openable]
      additionalProperties: false
      description: 代表の所在。GET /api/videos/{id} の応答にだけ入る
      properties:
        path:
          type: string
          description: 代表の所在の絶対パス。サーバーから見たパスで、コンテナ内ならコンテナ内のパス
        openable:
          type: boolean
          description: |
            要求元がループバックで、Host がループバックの名前で、サーバーが既定アプリを
            起動できる環境のとき true
...
    Progress:
      type: object
      required: [positionMs, completed, updatedAt]
      additionalProperties: false
      properties:
        positionMs:
          type: integer
          format: int64
        completed:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    Scan:
      type: object
      required: [id, state, total, completed, failed]
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        state:
          type: string
          enum: [running, done, failed]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        total:
          type: integer
          description: 変更なしを除いた取り込み対象ファイル数。対象の確定前は0
        completed:
          type: integer
          description: 取り込み処理に成功した対象ファイル数
        failed:
          type: integer
        error:
          type: string
          description: スキャン自体が失敗した理由
...
    Error:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
          description: >-
            機械可読なエラー種別。ここが正本で、Go の定数は生成物である
            （task generate）。新しい種別はまずここへ足す。
          enum:
            - not_found
            - invalid_request
            - internal
            - forbidden
            - conflict
            - invalid_media_directory
            - unsupported_media_directory
            - media_folder_not_found
            - overlapping_media_directories
            - scan_in_progress
            - media_folders_not_configured
            - directory_unavailable
            - probe_not_failed
            - open_unavailable
            - file_missing
        message:
          type: string
          description: 人が読むための説明。利用者にそのまま提示してよい文言にする